+++ crates/node/src/lib.rs
@@ -39,6 +39,7 @@ mod protocol;
 mod protocol_version;
 mod providers;
 pub mod requests;
+mod run;
 mod runtime;
 mod storage;
 pub mod tracing;
